+++ RSPiX/Src/GREEN/3D/render.cpp
@@ -31,7 +31,7 @@ typedef struct
 // rendered.
 // sX and sY are additional offsets into pimDst
 //
-void	DrawTri_ZColorFog(UCHAR* pDstOffset,long lDstP,
+void	DrawTri_ZColorFog(UCHAR* pDstOffset,int32_t lDstP,
 			RP3d* p1,RP3d* p2,RP3d* p3,
 			RZBuffer* pZB,UCHAR* pFog,
 			int16_t sOffsetX/* = 0*/,		// In: 2D offset for pZB.
@@ -166,16 +166,16 @@ void	DrawTri_ZColorFog(UCHAR* pDstOffset,long lDstP,
 	if (y2 + y3 == 0) return; // don;t bother drawing horiz line
 
 	// get relative floating point x coordinates: (32-bit differences)
-	long fx1 = pv1->x.val;
-	long fz1 = pv1->z.val;
+	int32_t fx1 = pv1->x.val;
+	int32_t fz1 = pv1->z.val;
 
-	long fx2 = pv2->x.val - fx1;
-	long fx3 = pv3->x.val - fx1;
-	long fz2 = pv2->z.val - fz1;
-	long fz3 = pv3->z.val - fz1;
+	int32_t fx2 = pv2->x.val - fx1;
+	int32_t fx3 = pv3->x.val - fx1;
+	int32_t fz2 = pv2->z.val - fz1;
+	int32_t fz3 = pv3->z.val - fz1;
 
 	// calculate the top two edge slopes with 32-bit accuracy:
-	long fx2inc,fz2inc;
+	int32_t fx2inc,fz2inc;
 	if (y2) 
 		{
 		fx2inc = fx2 / y2; // stuck with division using fx32!
@@ -186,8 +186,8 @@ void	DrawTri_ZColorFog(UCHAR* pDstOffset,long lDstP,
 		*/
 		}
 
-	long fx3inc = fx3 / y3; // stuck with division using fx32!
-	long fz3inc = fz3 / y3; // stuck with division using fx32!
+	int32_t fx3inc = fx3 / y3; // stuck with division using fx32!
+	int32_t fz3inc = fz3 / y3; // stuck with division using fx32!
 
 	// Set the two absolute edge positions
 	RFixedS32 x2,x3,z2,z3; 
@@ -198,11 +198,11 @@ void	DrawTri_ZColorFog(UCHAR* pDstOffset,long lDstP,
 	//TRACE("SBASE Zpt = %hd\n",pv1->z.mod);
 
 
-	long lP = lDstP;
+	int32_t lP = lDstP;
 	// add in extra piece uv rounding!
 	UCHAR* pDst = pDstOffset + lP * pv1->y.mod + pv1->x.mod + x2.mod; 
 	int16_t* pBufZ = pZB -> GetZPtr(pv1->x.mod + x2.mod + sOffsetX, pv1->y.mod + sOffsetY);
-	long lZP = pZB->m_lP; // in words!!!
+	int32_t lZP = pZB->m_lP; // in words!!!
 
 	// Draw the upper triangle! (Assuming fx2inc < fx3inc.....)
 	int16_t x,y;
@@ -224,7 +224,7 @@ void	DrawTri_ZColorFog(UCHAR* pDstOffset,long lDstP,
 
 	if (y2 == 0) // p1.y == p2.y
 		{
-		long fx1inc,fz1inc;
+		int32_t fx1inc,fz1inc;
 		RFixedS32 x1,z1; // Absolute positions
 
 		// Let point I be to the LEFT of point II:
@@ -269,7 +269,7 @@ void	DrawTri_ZColorFog(UCHAR* pDstOffset,long lDstP,
 			xdel = x2.mod - x1.mod;
 			//if (hzdel) hzinc.val = (z3.val - z2.val) / hzdel;
 			//***************8 flipped the inc value:!
-			if (xdel) fzinc.val = long(z2.mod - z1.mod) * RInitNum::OneOver[xdel];
+			if (xdel) fzinc.val = int32_t(z2.mod - z1.mod) * RInitNum::OneOver[xdel];
 			//if (hzdel) Mul(hzinc,z3.val - z2.val,CInitNum::OneOver[hzdel]);
 
 			// Assume 2 to 3:
@@ -351,7 +351,7 @@ void	DrawTri_ZColorFog(UCHAR* pDstOffset,long lDstP,
 				fz.mod += sBaseZ;
 				xdel = x3.mod - x2.mod;
 				//if (hzdel) hzinc.val = (z3.val - z2.val) / hzdel;
-				if (xdel) fzinc.val = long(z3.mod - z2.mod) * RInitNum::OneOver[xdel];
+				if (xdel) fzinc.val = int32_t(z3.mod - z2.mod) * RInitNum::OneOver[xdel];
 				//if (hzdel) Mul(hzinc,z3.val - z2.val,CInitNum::OneOver[hzdel]);
 
 				// Assume 2 to 3:
@@ -392,7 +392,7 @@ void	DrawTri_ZColorFog(UCHAR* pDstOffset,long lDstP,
 
 			xdel = x2.mod - x3.mod; //+ x to z
 			//if (hzdel) hzinc.val = (z2.val - z3.val) / hzdel;
-			if (xdel) fzinc.val = long(z2.mod - z3.mod) * RInitNum::OneOver[xdel];
+			if (xdel) fzinc.val = int32_t(z2.mod - z3.mod) * RInitNum::OneOver[xdel];
 			//if (hzdel) Mul(hzinc,z2.val - z3.val,CInitNum::OneOver[hzdel]);
 
 			// Assume 2 to 3:
@@ -475,7 +475,7 @@ void	DrawTri_wire(RImage* pimDst,int16_t sX,int16_t sY,
 // FLAT SHADED!
 // sX and sY are additional offsets into pimDst
 //
-void	DrawTri_ZColor(UCHAR* pDstOffset,long lDstP,
+void	DrawTri_ZColor(UCHAR* pDstOffset,int32_t lDstP,
 			RP3d* p1,RP3d* p2,RP3d* p3,
 			RZBuffer* pZB,UCHAR ucFlatColor,
 			int16_t sOffsetX/* = 0*/,		// In: 2D offset for pZB.
@@ -524,16 +524,16 @@ void	DrawTri_ZColor(UCHAR* pDstOffset,long lDstP,
 	if (y2 + y3 == 0) return; // don;t bother drawing horiz line
 
 	// get relative floating point x coordinates: (32-bit differences)
-	long fx1 = pv1->x.val;
-	long fz1 = pv1->z.val;
+	int32_t fx1 = pv1->x.val;
+	int32_t fz1 = pv1->z.val;
 
-	long fx2 = pv2->x.val - fx1;
-	long fx3 = pv3->x.val - fx1;
-	long fz2 = pv2->z.val - fz1;
-	long fz3 = pv3->z.val - fz1;
+	int32_t fx2 = pv2->x.val - fx1;
+	int32_t fx3 = pv3->x.val - fx1;
+	int32_t fz2 = pv2->z.val - fz1;
+	int32_t fz3 = pv3->z.val - fz1;
 
 	// calculate the top two edge slopes with 32-bit accuracy:
-	long fx2inc,fz2inc;
+	int32_t fx2inc,fz2inc;
 	if (y2) 
 		{
 		fx2inc = fx2 / y2; // stuck with division using fx32!
@@ -544,8 +544,8 @@ void	DrawTri_ZColor(UCHAR* pDstOffset,long lDstP,
 		*/
 		}
 
-	long fx3inc = fx3 / y3; // stuck with division using fx32!
-	long fz3inc = fz3 / y3; // stuck with division using fx32!
+	int32_t fx3inc = fx3 / y3; // stuck with division using fx32!
+	int32_t fz3inc = fz3 / y3; // stuck with division using fx32!
 
 	// Set the two absolute edge positions
 	RFixedS32 x2,x3,z2,z3; 
@@ -556,11 +556,11 @@ void	DrawTri_ZColor(UCHAR* pDstOffset,long lDstP,
 	//TRACE("SBASE Zpt = %hd\n",pv1->z.mod);
 
 
-	long lP = lDstP;
+	int32_t lP = lDstP;
 	// add in extra piece uv rounding!
 	UCHAR* pDst = pDstOffset + lP * pv1->y.mod + pv1->x.mod + x2.mod; 
 	int16_t* pBufZ = pZB -> GetZPtr(pv1->x.mod + x2.mod + sOffsetX, pv1->y.mod + sOffsetY);
-	long lZP = pZB->m_lP; // in words!!!
+	int32_t lZP = pZB->m_lP; // in words!!!
 
 	// Draw the upper triangle! (Assuming fx2inc < fx3inc.....)
 	int16_t x,y;
@@ -582,7 +582,7 @@ void	DrawTri_ZColor(UCHAR* pDstOffset,long lDstP,
 
 	if (y2 == 0) // p1.y == p2.y
 		{
-		long fx1inc,fz1inc;
+		int32_t fx1inc,fz1inc;
 		RFixedS32 x1,z1; // Absolute positions
 
 		// Let point I be to the LEFT of point II:
@@ -627,7 +627,7 @@ void	DrawTri_ZColor(UCHAR* pDstOffset,long lDstP,
 			xdel = x2.mod - x1.mod;
 			//if (hzdel) hzinc.val = (z3.val - z2.val) / hzdel;
 			//***************8 flipped the inc value:!
-			if (xdel) fzinc.val = long(z2.mod - z1.mod) * RInitNum::OneOver[xdel];
+			if (xdel) fzinc.val = int32_t(z2.mod - z1.mod) * RInitNum::OneOver[xdel];
 			//if (hzdel) Mul(hzinc,z3.val - z2.val,CInitNum::OneOver[hzdel]);
 
 			// Assume 2 to 3:
@@ -709,7 +709,7 @@ void	DrawTri_ZColor(UCHAR* pDstOffset,long lDstP,
 				fz.mod += sBaseZ;
 				xdel = x3.mod - x2.mod;
 				//if (hzdel) hzinc.val = (z3.val - z2.val) / hzdel;
-				if (xdel) fzinc.val = long(z3.mod - z2.mod) * RInitNum::OneOver[xdel];
+				if (xdel) fzinc.val = int32_t(z3.mod - z2.mod) * RInitNum::OneOver[xdel];
 				//if (hzdel) Mul(hzinc,z3.val - z2.val,CInitNum::OneOver[hzdel]);
 
 				// Assume 2 to 3:
@@ -750,7 +750,7 @@ void	DrawTri_ZColor(UCHAR* pDstOffset,long lDstP,
 
 			xdel = x2.mod - x3.mod; //+ x to z
 			//if (hzdel) hzinc.val = (z2.val - z3.val) / hzdel;
-			if (xdel) fzinc.val = long(z2.mod - z3.mod) * RInitNum::OneOver[xdel];
+			if (xdel) fzinc.val = int32_t(z2.mod - z3.mod) * RInitNum::OneOver[xdel];
 			//if (hzdel) Mul(hzinc,z2.val - z3.val,CInitNum::OneOver[hzdel]);
 
 			// Assume 2 to 3:
@@ -817,7 +817,7 @@ void	DrawTri_ZColor(UCHAR* pDstOffset,long lDstP,
 // sX and sY are additional offsets into pimDst
 // There is NO Z_BUFFER here!  It is JUST a polygon drawer
 //
-void	DrawTri(UCHAR* pDstOffset,long lDstP,
+void	DrawTri(UCHAR* pDstOffset,int32_t lDstP,
 			RP3d* p1,RP3d* p2,RP3d* p3,
 			UCHAR ucFlatColor)
 	{
@@ -861,13 +861,13 @@ void	DrawTri(UCHAR* pDstOffset,long lDstP,
 	if (y2 + y3 == 0) return; // don;t bother drawing horiz line
 
 	// get relative floating point x coordinates: (32-bit differences)
-	long fx1 = pv1->x.val;
+	int32_t fx1 = pv1->x.val;
 
-	long fx2 = pv2->x.val - fx1;
-	long fx3 = pv3->x.val - fx1;
+	int32_t fx2 = pv2->x.val - fx1;
+	int32_t fx3 = pv3->x.val - fx1;
 
 	// calculate the top two edge slopes with 32-bit accuracy:
-	long fx2inc;
+	int32_t fx2inc;
 	if (y2) 
 		{
 		fx2inc = fx2 / y2; // stuck with division using fx32!
@@ -876,13 +876,13 @@ void	DrawTri(UCHAR* pDstOffset,long lDstP,
 		*/
 		}
 
-	long fx3inc = fx3 / y3; // stuck with division using fx32!
+	int32_t fx3inc = fx3 / y3; // stuck with division using fx32!
 
 	// Set the two absolute edge positions
 	RFixedS32 x2,x3; 
 	x2.val = x3.val = pv1->x.frac;	// preserve floating point x!
 
-	long lP = lDstP;
+	int32_t lP = lDstP;
 	// add in extra piece uv rounding!
 	UCHAR* pDst = pDstOffset + lP * pv1->y.mod + pv1->x.mod + x2.mod; 
 
@@ -905,7 +905,7 @@ void	DrawTri(UCHAR* pDstOffset,long lDstP,
 
 	if (y2 == 0) // p1.y == p2.y
 		{
-		long fx1inc;
+		int32_t fx1inc;
 		RFixedS32 x1; // Absolute positions
 
 		// Let point I be to the LEFT of point II:
