+++ RSPiX/Src/GREEN/3D/render.h
@@ -45,7 +45,7 @@
 // Note that a unique fog table is needed for each
 // triangle color!
 //
-extern	void	DrawTri_ZColorFog(UCHAR* pDst,long lDstP,
+extern	void	DrawTri_ZColorFog(UCHAR* pDst,int32_t lDstP,
 			RP3d* p1,RP3d* p2,RP3d* p3,
 			RZBuffer* pZB,UCHAR* pFog, 
 			int16_t sOffsetX = 0,		// In: 2D offset for pZB.
@@ -56,7 +56,7 @@ extern	void	DrawTri_ZColorFog(UCHAR* pDst,long lDstP,
 extern	void	DrawTri_wire(RImage* pimDst,int16_t sX,int16_t sY,
 			RP3d* p1,RP3d* p2,RP3d* p3,UCHAR ucColor); 
 
-extern	void	DrawTri_ZColor(UCHAR* pDst,long lDstP,
+extern	void	DrawTri_ZColor(UCHAR* pDst,int32_t lDstP,
 			RP3d* p1,RP3d* p2,RP3d* p3,
 			RZBuffer* pZB,UCHAR pFlatColor,
 			int16_t sOffsetX = 0,		// In: 2D offset for pZB.
@@ -67,7 +67,7 @@ extern	void	DrawTri_ZColor(UCHAR* pDst,long lDstP,
 // sX and sY are additional offsets into pimDst
 // There is NO Z_BUFFER here!  It is JUST a polygon drawer
 //
-extern	void	DrawTri(UCHAR* pDstOffset,long lDstP,
+extern	void	DrawTri(UCHAR* pDstOffset,int32_t lDstP,
 			RP3d* p1,RP3d* p2,RP3d* p3,
 			UCHAR ucFlatColor);
