+++ RSPiX/Src/ORANGE/Attribute/attribute.cpp
@@ -266,7 +266,7 @@ int16_t RAttributeMap::AllocateMap(ULONG ulSize, ULONG ulDetailMapSize)
 //
 //////////////////////////////////////////////////////////////////////
 /*
-USHORT RAttributeMap::GetAttribute(long lX, long lY)
+USHORT RAttributeMap::GetAttribute(int32_t lX, int32_t lY)
 {
 	if ((lX/m_sScaleX) > m_lWidth || (lY/m_sScaleY) > m_lHeight || lX < 0 || lY < 0)
 		return ATTRIBUTE_NOT_WALKABLE;
@@ -279,7 +279,7 @@ USHORT RAttributeMap::GetAttribute(long lX, long lY)
 }
 */
 
-USHORT RAttributeMap::GetAttribute(long lX, long lY)
+USHORT RAttributeMap::GetAttribute(int32_t lX, int32_t lY)
 {
 	USHORT usBlockData;
 
@@ -340,13 +340,13 @@ USHORT RAttributeMap::GetAttribute(long lX, long lY)
 //////////////////////////////////////////////////////////////////////
 
 /*
-USHORT RAttributeMap::GetAttribute(long lTopCoord, long lBottomCoord,
-											 long lLeftCoord, long lRightCoord)
+USHORT RAttributeMap::GetAttribute(int32_t lTopCoord, int32_t lBottomCoord,
+											 int32_t lLeftCoord, int32_t lRightCoord)
 {
-	long lTop = lTopCoord / m_sScaleY;
-	long lBottom = lBottomCoord / m_sScaleY;
-	long lLeft = lLeftCoord / m_sScaleX;
-	long lRight = lRightCoord / m_sScaleX;
+	int32_t lTop = lTopCoord / m_sScaleY;
+	int32_t lBottom = lBottomCoord / m_sScaleY;
+	int32_t lLeft = lLeftCoord / m_sScaleX;
+	int32_t lRight = lRightCoord / m_sScaleX;
 
 	// Clip the box to the map size if necessary
 	if (lTop >= m_lHeight)
@@ -366,8 +366,8 @@ USHORT RAttributeMap::GetAttribute(long lTopCoord, long lBottomCoord,
 	if (lRight < 0)
 		lRight = 0;
 
-	long lRow;
-	long lCol;
+	int32_t lRow;
+	int32_t lCol;
 	USHORT usResult = 0;
 	USHORT usAttrib = 0;
 	USHORT usMin = m_pusMap[lTop*m_lWidth + lLeft];
@@ -393,13 +393,13 @@ USHORT RAttributeMap::GetAttribute(long lTopCoord, long lBottomCoord,
 }
 */
 
-USHORT RAttributeMap::GetAttribute(long lTopCoord, long lBottomCoord,
-											 long lLeftCoord, long lRightCoord)
+USHORT RAttributeMap::GetAttribute(int32_t lTopCoord, int32_t lBottomCoord,
+											 int32_t lLeftCoord, int32_t lRightCoord)
 {
-	long lTop = lTopCoord;// / m_sScaleY;
-	long lBottom = lBottomCoord;// / m_sScaleY;
-	long lLeft = lLeftCoord;// / m_sScaleX;
-	long lRight = lRightCoord;// / m_sScaleX;
+	int32_t lTop = lTopCoord;// / m_sScaleY;
+	int32_t lBottom = lBottomCoord;// / m_sScaleY;
+	int32_t lLeft = lLeftCoord;// / m_sScaleX;
+	int32_t lRight = lRightCoord;// / m_sScaleX;
 
 	// Clip the box to the map size if necessary
 	if (lTop >= m_lWorldHeight)
@@ -419,8 +419,8 @@ USHORT RAttributeMap::GetAttribute(long lTopCoord, long lBottomCoord,
 	if (lRight < 0)
 		lRight = 0;
 
-	long lRow;
-	long lCol;
+	int32_t lRow;
+	int32_t lCol;
 	USHORT usResult = 0;
 	USHORT usAttrib = 0;
 	USHORT usFlags = 0;
@@ -527,7 +527,7 @@ int16_t RAttributeMap::Load(RFile* prf)
 										{
 											if (AllocateMap(m_lWidth * m_lHeight, m_sNumDetailMaps * m_sScaleX * m_sScaleY) == SUCCESS)
 											{
-												if (prf->Read(m_pusMap, m_lWidth*m_lHeight) == (long) (m_lWidth*m_lHeight))
+												if (prf->Read(m_pusMap, m_lWidth*m_lHeight) == (int32_t) (m_lWidth*m_lHeight))
 												{
 													if (prf->Read(m_pusDetailMap, m_sNumDetailMaps*m_sScaleX*m_sScaleY) == m_sNumDetailMaps*m_sScaleX*m_sScaleY)
 													{
