+++ RSPiX/Src/ORANGE/Attribute/attribute.h
@@ -79,10 +79,10 @@
 class RAttributeMap
 {
 	public:
-		long m_lWidth;
-		long m_lHeight;
-		long m_lWorldWidth;
-		long m_lWorldHeight;
+		int32_t m_lWidth;
+		int32_t m_lHeight;
+		int32_t m_lWorldWidth;
+		int32_t m_lWorldHeight;
 		int16_t m_sScaleX;
 		int16_t m_sScaleY;
 		int16_t m_sNumDetailMaps;
@@ -104,10 +104,10 @@ class RAttributeMap
 		int16_t Load(RFile* prf);
 
 		// Single point attribute
-		USHORT GetAttribute(long lX, long lY);
+		USHORT GetAttribute(int32_t lX, int32_t lY);
 
 		// Rectangle attribute
-		USHORT GetAttribute(long lTop, long lBottom, long lLeft, long lRight);
+		USHORT GetAttribute(int32_t lTop, int32_t lBottom, int32_t lLeft, int32_t lRight);
 
 		// Get just the low 8 bits (flags) from the last GetAttribute Call
 		UCHAR GetFlags()
